Safari-611.1.10/LayoutTests/http/tests/notifications/request-in-detached-frame.html

<!DOCTYPE html>
<script src="/resources/js-test-pre.js"></script>
<script src="/resources/notifications-test-pre.js"></script>
<p id="description"></p>
<div id="console"></div>
<iframe id="testFrame" src="resources/request-in-detached-frame-subframe.html"></iframe>
<script>
description("This test checks accessing Notification.permission in a detached iframe.");

if (window.testRunner)
    testRunner.waitUntilDone();

onload = () => {
    testFrame.contentWindow.postMessage("foo", "*");
    setTimeout(() => {
        testCompleted();
    }, 100);
};
</script>
