LayoutTests/http/tests/notifications/request-in-detached-frame-expected.txt

This test checks accessing Notification.permission in a detached iframe.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Notification.permission is "default"
PASS successfullyParsed is true

TEST COMPLETE
